requirements.txt
moviepy
speechrecognition
pydub
requests
anthropic
python-dotenv